+++ .gitignore
@@ -1,1 +1,1 @@
-./_pycache_
+**/__pycache__
