+++ src/.env
@@ -1,0 +1,10 @@
+HEADER=1024
+FORMAT=utf-8
+
+DISCONNECT_MESSAGE=!DISCONNECT
+USER_LIST_UPDATE=!USER_LIST
+USERNAME_TAKEN=Taken
+USERNAME_ACCEPTED=Ok
+
+WHISPER_CMD=/w
+DM_CMD=/dm
